python_flask_api/flask-jwt-extended/code/requirements.txt
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-extended